requirements.txt
astroid==2.4.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-logging==2.2.0
google-cloud-secret-manager==2.2.0
google-oauth==1.0.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
httplib2==0.19.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
libcst==0.3.17
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
